rename yodel-process-end-marker to yodel--process-end-text
It's not a marker object, even though we're treating it like one.
It shouldn't be customizable, either.
Eventually I'd like to replace this by inserting a proper marker
via the filter function, but I have to find out if there's a reliable
way to detect the appropriate batch of subprocess output during the filter.