yodel: Add lexical-binding workaround
We want to ensure the subprocess is run with lexical binding enabled.
Instead of relying on the end-user to enable lexical binding,
we manually eval. It's "uglier", but it should be guaranteed to work.

I ran into this because even when outputting a report using Org scr blocks
with :lexical t specified, there is no way to detect whether yodel is being run
in a lexical environment. Org babel uses `eval' as well (with lexical binding
enabled), but since the evaluation is associated with the current buffer
(in this case an Org mode buffer), the file variable `lexical-binding' is nil.
I proposed a patch to remedy this, but it has not been taken up as of yet.

https://orgmode.org/list/87zgskcvgn.fsf@gmail.com/T/#t