yodel-parallel: append all args
We were mistakenly constructing the common arg list under the assumption that it
was a proper plist. However, it is a plist*, so variadic keywords would only
have their first argument added.e.g.

;;we lose two here
(plist-get '(:variadic* one two) :variadic*) ; one

Since we only need the :user-dir argument from
each test case, we can access that as if args were a plist and append the
remaining args for each yodel form.